move dep-installer stuff to own dir
