+++ scripts/binary/smoke.js
@@ -7,6 +7,7 @@ const Promise = require('bluebird')
 const os = require('os')
 const verify = require('../../cli/lib/tasks/verify')
 const Fixtures = require('@tooling/system-tests/lib/fixtures')
+const { scaffoldCommonNodeModules } = require('@tooling/system-tests/lib/dep-installer')
 
 const fs = Promise.promisifyAll(fse)
 
@@ -160,7 +161,7 @@ const runFailingProjectTest = function (buildAppExecutable, e2e) {
 }
 
 const test = async function (buildAppExecutable) {
-  await Fixtures.scaffoldCommonNodeModules()
+  await scaffoldCommonNodeModules()
   Fixtures.scaffoldProject('e2e')
   const e2e = Fixtures.projectPath('e2e')
