+++ system-tests/test/non_root_read_only_fs_spec.ts
@@ -2,6 +2,7 @@ import * as fs from 'fs'
 import * as path from 'path'
 import systemTests from '../lib/system-tests'
 import Fixtures from '../lib/fixtures'
+import { scaffoldCommonNodeModules } from '@tooling/system-tests/lib/dep-installer'
 
 describe('e2e readonly fs', function () {
   systemTests.setup()
@@ -24,7 +25,7 @@ describe('e2e readonly fs', function () {
 
   const onRun = async (exec) => {
     Fixtures.scaffoldProject('read-only-project-root')
-    await Fixtures.scaffoldCommonNodeModules()
+    await scaffoldCommonNodeModules()
     chmodr(projectPath, 0o500)
 
     await exec()
